Add FFI library class for miniaudio DLL
Also add classes for miniaudio enums.

The callout for ma_version_string() works as expected. The callout for ma_encoder_config_init(), however, crashes the image. I suspect that's because the returned struct, which is converted to an FFIOpaqueObject (because miniaudio doesn't guarantee ABI compatibility across releases or platforms), needs to be a pointer to an object on the heap, but is instead a struct allocated on the stack.